Add tear down flag

+++ tests/__init__.py
@@ -13,6 +13,8 @@ from workflow_array_ephys.paths import get_ephys_root_data_dir
 
 # ------------------- SOME CONSTANTS -------------------
 
+_tear_down = False
+
 test_user_data_dir = pathlib.Path('./tests/user_data')
 test_user_data_dir.mkdir(exist_ok=True)
 
@@ -84,7 +86,8 @@ def pipeline():
            'session': pipeline.session,
            'get_ephys_root_data_dir': pipeline.get_ephys_root_data_dir}
 
-    pipeline.subject.Subject.delete()
+    if _tear_down:
+        pipeline.subject.Subject.delete()
 
 
 @pytest.fixture
@@ -196,7 +199,8 @@ def kilosort_paramset(pipeline):
 
     yield params_ks
 
-    (ephys.ClusteringParamSet & 'paramset_idx = 0').delete()
+    if _tear_down:
+        (ephys.ClusteringParamSet & 'paramset_idx = 0').delete()
 
 
 @pytest.fixture
@@ -207,7 +211,8 @@ def ephys_recordings(pipeline, ingest_sessions):
 
     yield
 
-    ephys.EphysRecording.delete()
+    if _tear_down:
+        ephys.EphysRecording.delete()
 
 
 @pytest.fixture
@@ -228,7 +233,8 @@ def clustering_tasks(pipeline, kilosort_paramset, ephys_recordings):
 
     yield
 
-    ephys.ClusteringTask.delete()
+    if _tear_down:
+        ephys.ClusteringTask.delete()
 
 
 @pytest.fixture
@@ -239,7 +245,8 @@ def clustering(clustering_tasks, pipeline):
 
     yield
 
-    ephys.Clustering.delete()
+    if _tear_down:
+        ephys.Clustering.delete()
 
 
 @pytest.fixture
@@ -251,4 +258,5 @@ def curations(clustering, pipeline):
 
     yield
 
-    ephys.Curation.delete()
+    if _tear_down:
+        ephys.Curation.delete()
